refactored lookupItems to Item Controller instead of OrderController
